Moved SAS data to the ECU, added SAS filtering.

+++ SAS.X/ADC.h
@@ -19,7 +19,7 @@ typedef enum{
 
 void initADC(void);
 void FilterADC();
-float GetADC(value RequestValue);
+unsigned int GetADC(value RequestValue);
 //bool ADCFinish();
 
 #endif	/* ADC_H */
